// src/app/models/setting.ts
export class Setting {
  currency: string;
  currency_symbol: string;
  properties_limit: string;
  cards_limit: string;
  bank_accounts_limit: string;
  become_a_provider_url: string;
  become_a_manager_url: string;
  default_designer_email: string;
  default_buyer_email: string;
  terms_and_condition_url: string;
  privacy_policy_url: string;
  stripe_test_payment: string;
  updated_at: string;
}